Solve the last problem
But still there are some other problems, like setting the FFT window size to large number is set to size smaller then the preferred size. And also have to have better accuracy for the frequencies in tooltips and when using the time draw panel, the tooltip breaks for certain values.