Fix permission bug and JFileChooser name saving
When we were creating new audio file in a place where we didn't had permissions, then we got exception.
Between the calls of the synthesizers recording, we didn't save the filepath between calls (that means that we changed the file path, then we wanted to change it again it switched to the default place in the documents), this is fixed now.