Add things with todoMark, still finding solution
The problem is that when the scrollbar is made larger when zooming (or smaller when unzooming) there occurs shift in the wave, so for a brief moment the scrollbar is at the old location, but for the new size of scrollbar, so it shows at wrong place, which is at 1/2 or 2 times the supposed location.
There are 3 solutions now.
First one is just fix the marking, which will be fixed as I did it now, but just need to find the correct currentScroll, which will be after the zoom/unzoom and based on that change it so it will be the same marking as after the zoom.
Second is to just save image before zoom, and show it as bridge, until it could be drawn correctly. The code is also there, just need to correctly pass the wave to the image.
Third solution, which is the best one, but I kind of don't want to do is: When zooming, don't shift the wave buffer, keep it as it is and shift it only when the both enlargement and shift of the scrollbar was done.