Fix JAR file made by ANT issues
The issue was just in the way the plugins are read, first we had relative path to the jar file instead of absolute and the whole algorithm of reading the synth plugins was too complicated and wrong.

There is still one, when launching the jar using right click, it doesn't seem to know the location of the libraries, when calling it from command line from the directory of the build.xml all is fine. Also moving the libs directory to the same directory as the jar or inside the jar doesn't help, but it won't be really that difficult issue to solve.