Rename: JPanelWithMovableJPanels -> DiagramPanel
And also rename mainPanel -> diagramPanel in case of variables and arguments

DiagramPanel says enough info - Diagram says that it has to do something with diagram and panel says that it is GUI component.
(colliding a bit with the DiagramJSplitPane, but it isn't too bad - and naming the other one DiagramBottomPanel is kind of weird).