Rename: AudioWavePanelEverything -> WaveMainPanel
Also replace: awpe -> waveMainPanel

I couldn't think of better name to describe it, it had to contain wave and panel strings, then we needed to describe that it is the main class, that it contains everything. Word control really doesn't work, since it isn't controlling anything, wrapper is misleading, WithEverything sounds weird and doesn't say that much. The only issue is that it collides a bit with WavePanel, but we can't really fix that, we could use name TrackMainPanel or just TrackPanel, but then it would look like it doesn't have anything to do with the other waves classes and also track is very similar in meaning to wave.