Add copyInternalState methods
Also add some comments because of that.

And change some things around that because of future proofing - for example instead of rewriting the copyPanel method we just solve the additional copying in the copy constructor. It usually involves the call of copyInternalState. We didn't put the method to the code of copyPanel method because it doesn't work for filters. Here is the full comment:
Thought: Maybe I could call it at the end of the copyPanel method, but it could a bit problematic with some additional generators - for example when you copy filter - you copy the values which are currently there - and the values may be in non-consistent state, because all the values are changing since we may be currently in process of generating samples. Because of that it is better to just have all the values set to default for the target of copy (this instance) - like no sample was played before.