Rename package: NoPhase -> nophase
I don't know about this, also I am not sure about the generator names. But I guess that it has to be like that, I can't have the phase and non-phase version with the same names, then if I want to use both in class I have to reference one of the by full name and also I wouldn't know which is which. But maybe I will switch it so the ones with phase won't have phase in name, but the ones without will be named like SineGeneratorNoPhase or SineGeneratorSimple, but it doesn't contain the information about phase, so either the NoPhase variant or keep it like it is.