Rename file: PluginDefaultIFace -> PluginBaseIFace
Both names are fine, but I think that when it contains default it is a bit worse, since we are often referencing to default panel (when we are talking about the panel from annotations). Also it is shorter. ... Note now that I think about it I will probably reference to such panel as annotation panel instead of default panel.