Fix the jar plugin loading (load only 1 dir above)
Before I was loading whatever came first, but I could have all the directories and sometime in the future all the units (talking about the order in the jar), that would result in having all the units in 1 level, instead in the demanded hierarchy. So now we go to recursion only if it is the next subdirectory. It still isn't perfect because if I have multiple units on 1 level, then they can be in between the "directories", but I am not fixing that now, or maybe I am it is easy fix.