Clean-up: test package
But I don't know, it just really isn't something to be proud of. It does what it has to but not in some shining way.
Also I have to let intellij fix the indentation etc for consistency, don't really have time to do it by hand, that means either changing code style to my new one, or keep the code style as is.